pyproject.toml
[project]
name = "abca4-faf"
version = "0.1.0"
requires-python = ">=3.13"

dependencies = [
    "aiosqlite==0.22.1",
    "alembic==1.18.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "apprise==1.9.6",
    "asgi-lifespan==2.1.0",
    "asyncpg==0.31.0",
    "attrs==25.4.0",
    "beartype==0.22.9",
    "brisque==0.0.17",
    "cachetools==6.2.4",
    "cairocffi==1.7.1",
    "cairosvg==2.8.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "coolname==2.2.0",
    "cryptography==46.0.3",
    "cssselect2==0.8.0",
    "cycler==0.12.1",
    "dask==2025.10.0",
    "dateparser==1.2.2",
    "defusedxml==0.7.1",
    "distributed==2025.10.0",
    "docker==7.1.0",
    "dotenv==0.9.9",
    "exceptiongroup==1.3.1",
    "fakeredis==2.33.0",
    "fastapi==0.128.0",
    "fonttools==4.60.1",
    "fsspec==2025.9.0",
    "graphviz==0.21",
    "greenlet==3.3.0",
    "griffe==1.15.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "humanize==4.15.0",
    "hyperframe==6.1.0",
    "idna==3.11",
    "image-quality==1.2.7",
    "imagecodecs==2025.8.2",
    "imageio==2.37.0",
    "imagesize==1.4.1",
    "importlib-metadata==8.7.1",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "jinja2-humanize-extension==0.4.0",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "kiwisolver==1.4.9",
    "lazy-loader==0.4",
    "libsvm==3.23.0.4",
    "libsvm-official==3.36.0",
    "locket==1.0.0",
    "lupa==2.6",
    "lxml==6.0.2",
    "mako==1.3.10",
    "markdown==3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "matplotlib==3.10.7",
    "mdurl==0.1.2",
    "msgpack==1.1.2",
    "networkx==3.5",
    "numpy==2.2.6",
    "oauthlib==3.3.1",
    "opencv-python==4.12.0.88",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-prometheus==0.60b1",
    "opentelemetry-instrumentation==0.60b1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "orjson==3.11.5",
    "packaging==25.0",
    "pandas==2.3.3",
    "partd==1.4.2",
    "pathspec==1.0.3",
    "peewee==3.18.2",
    "peewee-enum-field==0.2.0",
    "pillow==12.0.0",
    "playhouse==0.1.0",
    "pluggy==1.6.0",
    "prefect==3.6.10",
    "prometheus-client==0.24.1",
    "psutil==7.1.2",
    "py-key-value-aio==0.3.0",
    "py-key-value-shared==0.3.0",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.11.0",
    "pydantic-settings==2.12.0",
    "pydocket==0.16.6",
    "pygments==2.19.2",
    "pymysql==1.1.2",
    "pyparsing==3.2.5",
    "pytest==9.0.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-json-logger==4.0.0",
    "python-pptx==1.0.2",
    "python-slugify==8.0.4",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "readchar==4.2.1",
    "redis==7.1.0",
    "referencing==0.37.0",
    "regex==2026.1.14",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rfc3339-validator==0.1.4",
    "rich==14.2.0",
    "rpds-py==0.30.0",
    "ruamel-yaml==0.19.1",
    "ruamel-yaml-clib==0.2.15",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.2",
    "scikits-bootstrap==1.1.1",
    "scipy==1.16.2",
    "semver==3.0.4",
    "shellingham==1.5.4",
    "shortcuts==0.11.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlalchemy==2.0.45",
    "starlette==0.50.0",
    "svgwrite==1.4.3",
    "tblib==3.2.0",
    "text-unidecode==1.3",
    "threadpoolctl==3.6.0",
    "tifffile==2025.10.16",
    "tinycss2==1.4.0",
    "toml==0.10.2",
    "toolz==1.1.0",
    "tornado==6.5.2",
    "typer==0.20.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3==2.5.0",
    "uvicorn==0.40.0",
    "webencodings==0.5.1",
    "websockets==15.0.1",
    "whenever==0.9.5",
    "wrapt==1.17.3",
    "xlsxwriter==3.2.9",
    "zict==3.0.0",
    "zipp==3.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[project.scripts]
pipeline = "faf28_workflows.bootstrap:main"
